requirements.txt
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
flake8==7.3.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
numpy==2.0.2
packaging==25.0
pandas==2.3.0
pandas-stubs==2.2.2.240807
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.14.0
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
tomli==2.2.1
types-python-dateutil==2.9.0.20250516
types-pytz==2025.2.0.20250516
types-requests==2.32.4.20250611
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
